refactor: extract HotkeySupportClasses.swift from HotkeyManager.swift
Extracts 2 @MainActor support classes (HotkeySettingsBinder,
HotkeyActionDispatcher) from HotkeyManager.swift into a new
HotkeySupportClasses.swift.

Byte-identical move; SHA verified against origin/main.

Applied memory rule "watch for annotations on line immediately preceding
deleted blocks": both @MainActor annotations (lines 5 and 48) belong to their
following classes and are included in the extraction range. HotkeyManager
class at line 88 is unaffected.

Reduces HotkeyManager.swift from 212 to 130 lines.

Tests: 667 (unchanged).
